Update and rename BIshop.java to Bishop.java
